+++ internal/provider/notification_channel/resource.go
@@ -22,7 +22,8 @@ func Resource() *schema.Resource {
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},
-		Schema: NotificationChannelResourceSchema(),
+		Schema:        NotificationChannelResourceSchema(),
+		CustomizeDiff: ValidateWebhookAuth,
 	}
 }